+++ app/controllers/activities_controller.rb
@@ -10,6 +10,9 @@ class ActivitiesController < BaseController
 
   def index
     @organisation = Organisation.find(organisation_id)
+
+    @deleted_activity_roda_identifier = params[:deleted_activity_roda_identifier]
+
     if @organisation.service_owner?
       @partner_organisations = Organisation.partner_organisations
 
@@ -63,6 +66,27 @@ class ActivitiesController < BaseController
         response.headers["Content-Disposition"] = "attachment; filename=\"#{@activity.transparency_identifier}.xml\""
       end
     end
+  end
+
+  def confirm_destroy
+    @activity = Activity.find(id)
+
+    authorize @activity, :destroy?
+
+    prepare_default_activity_trail(@activity)
+  end
+
+  def destroy
+    @activity = Activity.find(id)
+
+    authorize @activity, :destroy?
+
+    @activity.destroy
+
+    redirect_to organisation_activities_path(
+      @activity.organisation,
+      deleted_activity_roda_identifier: @activity.roda_identifier
+    )
   end
 
   def historic