+++ spec/controllers/activities_controller_spec.rb
@@ -175,4 +175,70 @@ RSpec.describe ActivitiesController do
       end
     end
   end
+
+  context "deleting activities" do
+    let(:activity) { create(:programme_activity, title: nil, form_state: "purpose") }
+    let(:policy) { instance_double(ActivityPolicy) }
+    let(:user) { instance_double(User) }
+
+    before do
+      allow(controller).to receive(:current_user).and_return(user)
+      allow(ActivityPolicy).to receive(:new).and_return(policy)
+    end
+
+    describe "#confirm_destroy" do
+      context "when authorised" do
+        it "responds with status 200 OK" do
+          allow(policy).to receive(:destroy?).and_return(true)
+          allow(user).to receive(:service_owner?).and_return(true)
+
+          get :confirm_destroy, params: {organisation_id: activity.organisation, activity_id: activity.id}
+
+          expect(response).to have_http_status(:ok)
+        end
+      end
+
+      context "when unauthorised" do
+        it "responds with status 401 Unauthorized" do
+          allow(policy).to receive(:destroy?).and_return(false)
+          allow(user).to receive(:service_owner?).and_return(false)
+
+          get :confirm_destroy, params: {organisation_id: activity.organisation, activity_id: activity.id}
+
+          expect(response).to have_http_status(:unauthorized)
+        end
+      end
+    end
+
+    describe "#destroy" do
+      context "when authorised" do
+        it "destroys the activity and redirects to the organisation activities path" do
+          allow(policy).to receive(:destroy?).and_return(true)
+
+          delete :destroy, params: {organisation_id: activity.organisation, id: activity.id}
+
+          expect(Activity.find_by(id: activity.id)).to be_nil
+
+          expect(response).to redirect_to(
+            organisation_activities_path(
+              activity.organisation,
+              deleted_activity_roda_identifier: activity.roda_identifier
+            )
+          )
+
+          expect(response).to have_http_status(:found)
+        end
+      end
+
+      context "when unauthorised" do
+        it "responds with status 401 Unauthorized" do
+          allow(policy).to receive(:destroy?).and_return(false)
+
+          delete :destroy, params: {organisation_id: activity.organisation, id: activity.id}
+
+          expect(response).to have_http_status(:unauthorized)
+        end
+      end
+    end
+  end
 end