+++ spec/policies/activity_policy_spec.rb
@@ -81,12 +81,18 @@ RSpec.describe ActivityPolicy do
 
         it { is_expected.to forbid_action(:update_linked_activity) }
       end
+
+      context "when the activity is untitled" do
+        let(:activity) { create(:programme_activity, title: "", form_state: "purpose") }
+
+        it { is_expected.to permit_action(:destroy) }
+      end
     end
 
     context "when the activity is a project" do
       let(:activity) { create(:project_activity) }
 
-      it "only permits download, show, and redact_from_iati" do
+      it "only permits show, redact_from_iati, download and set_commitment" do
         is_expected.to permit_action(:show)
         is_expected.to permit_action(:redact_from_iati)
         is_expected.to permit_action(:download)
@@ -132,12 +138,18 @@ RSpec.describe ActivityPolicy do
           end
         end
       end
+
+      context "when the activity is untitled" do
+        let(:activity) { create(:project_activity, title: "", form_state: "purpose") }
+
+        it { is_expected.to permit_action(:destroy) }
+      end
     end
 
     context "when the activity is a third-party project" do
       let(:activity) { create(:third_party_project_activity) }
 
-      it "only permits download, show, and redact_from_iati" do
+      it "only permits show, redact_from_iati, download and set_commitment" do
         is_expected.to permit_action(:show)
         is_expected.to permit_action(:redact_from_iati)
         is_expected.to permit_action(:download)
@@ -173,6 +185,12 @@ RSpec.describe ActivityPolicy do
             it { is_expected.to permit_action(:update_linked_activity) }
           end
         end
+      end
+
+      context "when the activity is untitled" do
+        let(:activity) { create(:third_party_project_activity, title: "", form_state: "purpose") }
+
+        it { is_expected.to permit_action(:destroy) }
       end
     end
   end